[#3] Refactor: reducer 함수 모듈화
